+++ CHANGELOG.md
@@ -8,4 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`f4208b3...main`][f4208b3...main].
 
+### Added
+
+- Added `ValueGenerator` ([#1]), by [@localheinz]
+
 [f4208b3...main]: https://github.com/ergebnis/data-generator/compare/f4208b3...main
+
+[#1]: https://github.com/ergebnis/data-generator/pull/1
+
+[@localheinz]: https://github.com/localheinz